+++ modeshape-common/src/main/resources/org/modeshape/common/CommonI18n.properties
@@ -109,3 +109,4 @@ log4jAvailable = Log4j located in the classpath. It will be used by ModeShape fo
 jdkFallback = No custom logger, SLF4J implementation or Log4j located in the classpath. ModeShape will use the JDK logger for logging.
 
 errorWhileClosingRingBufferConsumer = Error while closing the ring buffer consumer {0}: {1}
+errorClosingWrappedStream = Error while attempting to close a wrapped input stream
