+++ modeshape-jcr/src/main/java/org/modeshape/jcr/value/binary/InMemoryBinaryValue.java
@@ -65,11 +65,6 @@ public class InMemoryBinaryValue extends AbstractBinary {
     }
 
     @Override
-    public InputStream getStream() {
-        return new ByteArrayInputStream(this.bytes);
-    }
-
-    @Override
     public String getMimeType() throws IOException, RepositoryException {
         if (mimeType == null) {
             mimeType = store.getMimeType(this, null);
@@ -83,5 +78,10 @@ public class InMemoryBinaryValue extends AbstractBinary {
             mimeType = store.getMimeType(this, name);
         }
         return mimeType;
+    }
+
+    @Override
+    protected InputStream internalStream() {
+        return new ByteArrayInputStream(this.bytes);
     }
 }
